+++ src/BLE_Setup.cpp
@@ -14,7 +14,8 @@
 void setupBLE() {  // Common BLE setup for both client and server
   SS2K_LOG(BLE_SETUP_LOG_TAG, "Starting Arduino BLE Client application...");
   NimBLEDevice::init(userConfig->getDeviceName());
-  NimBLEDevice::setMTU(515);  //-- enabling this is very important for BLE firmware updates.
+  NimBLEDevice::setOwnAddrType(BLE_ADDR_PUBLIC);  // Garmin requires a public (non-random) MAC address to discover sensors
+  NimBLEDevice::setMTU(515);                      //-- enabling this is very important for BLE firmware updates.
   spinBLEClient.start();
   startBLEServer();
   SS2K_LOG(BLE_SETUP_LOG_TAG, "%s %s %s", userConfig->getConnectedPowerMeter(), userConfig->getConnectedHeartMonitor(), userConfig->getConnectedRemote());